LayoutTests/platform/glib/imported/w3c/web-platform-tests/css/css-text/text-fit/text-autospace-expected.txt

高輪Gateway
高輪Gateway
高輪Gateway
高輪Gateway

FAIL "text-autospace" and "consistent" work fine assert_approx_equals: The last letter aligns with the right edge expected 399 +/- 1 but got 99
FAIL "text-autospace" and "per-line-all" work fine assert_approx_equals: The last letter aligns with the right edge expected 399 +/- 1 but got 99
